Add IP dropdown and port validation
The default behavior was the attach to 0.0.0.0 and not allow the port to be edited when initializing. This change allows the server to start up binding to all interfaces (previous behavior) on 3000, but when stopped it allows the user to select a specific IP to bind to.

Port validation was added to ensure they're not entering garbage input.